// actions/create-card/schema.ts
import { z } from "zod";

//zod schema /w validations
export const CreateCard = z.object({
    title: z.string({
        required_error: "Title is required",
        invalid_type_error: "Title is required"
    }).min(2, {
        message: "Title is too short"
    }),
    boardId: z.string(),
    listId: z.string()
});